requirements.txt
pandas
numpy
requests
websocket-client
python-dotenv
pyyaml
pyarrow
tabulate
flask
